requirements.txt
python-telegram-bot>=20.0
pytz
asyncio
requests==2.28.1 